sewing-machine-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.5.0
python-dotenv==1.0.0
pydantic[email]==2.4.2
python-multipart==0.0.6
bcrypt==4.0.1
passlib==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0
xgboost==2.0.3
numpy==1.24.3
scikit-learn==1.4.2
joblib==1.3.2
requests==2.31.0